Add some documentation for cycled scanner handlers
This helps to try explaining how / when the various internal handlers
should be used. This is important to help developers understand various
threading flow orders, race conditions, and resource contention
problems. As part of the documentation this adds `MainThread` and
`WorkerThread` annotations to the deferred `Runnable` tasks' `run`
methods.

See Also:

- https://developer.android.com/studio/write/annotations.html#thread-annotations